event-driven-architecture/monolito/src/propiedadalpes/modulos/companias/aplicacion/__init__.py
from pydispatch import dispatcher

from .handlers import HandlerCompaniaIntegracion

from propiedadalpes.modulos.companias.dominio.eventos import CompaniaIngestada

dispatcher.connect(HandlerCompaniaIntegracion.handle_compania_ingestada, signal=f'{CompaniaIngestada.__name__}Integracion')
